django/insta_pjt/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==4.1
django-appconf==1.0.5
django-extensions==3.2.0
django-imagekit==4.1.0
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.4.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
traitlets==5.3.0
tzdata==2022.2
wcwidth==0.2.5
